+++ invisibleroads_macros_configuration/__init__.py
@@ -2,6 +2,7 @@ from .constants import SECRET_LENGTH
 from .settings import (
     Settings,
     fill_environment_variables,
+    fill_extensions,
     fill_secrets,
     set_default)
 from .formats import load_json