+++ invisibleroads_macros_configuration/settings.py
@@ -1,3 +1,4 @@
+from importlib import import_module
 from invisibleroads_macros_log import get_log
 from invisibleroads_macros_security import make_random_string
 from os.path import expandvars
@@ -29,13 +30,34 @@ def set_default(settings, key, default, parse=None):
 
 
 def fill_environment_variables(settings):
-    for k, v in settings.items():
-        if not isinstance(v, str):
-            continue
-        settings[k] = expandvars(v)
+    return {
+        k: expandvars(v)
+        if isinstance(v, str) else v
+        for k, v in settings.items()
+    }
 
 
 def fill_secrets(settings, secret_length=SECRET_LENGTH):
-    for k, v in settings.items():
-        if k.endswith('.secret') and not v:
-            settings[k] = make_random_string(secret_length)
+    return {
+        k: make_random_string(secret_length)
+        if k.endswith('.secret') and not v else v
+        for k, v in settings.items()
+    }
+
+
+def fill_extensions(settings):
+
+    def load_extensions(extension_specs):
+        extensions = []
+        for extension_spec in extension_specs:
+            module_spec, extension_name = extension_spec.rsplit(
+                '.', maxsplit=1)
+            module = import_module(module_spec)
+            extensions.append(getattr(module, extension_name))
+        return extensions
+
+    return {
+        k: load_extensions(v.split())
+        if k.endswith('.extensions') else v
+        for k, v in settings.items()
+    }