+++ CHANGES.md
@@ -1,3 +1,3 @@
 # 1.0
-- Add fill_environment_variables, fill_secrets
-- Restore set_default
+- Add fill_environment_variables, fill_extensions, fill_secrets
+- Restore Settings, set_default